Added Dependencies
Multiple Apps in 1 project. Will be the base for the overall project

Added the post repository to fetch posts through a extendable module.

Made a separate model for the incoming which emits the data from the Repository Posts

Added Weather Tests. Start for Posts App

Bloc Pattern with 2 pages

New initial commit